Update entity files.
