+++ deepwell/src/models/file_revision.rs
@@ -1,48 +1,44 @@
 //! SeaORM Entity. Generated by sea-orm-codegen 0.6.0
 
-use super::sea_orm_active_enums::PageRevisionType;
+use super::sea_orm_active_enums::FileRevisionType;
 use sea_orm::entity::prelude::*;
 use serde::{Deserialize, Serialize};
 
 #[derive(Clone, Debug, PartialEq, DeriveEntityModel, Serialize, Deserialize)]
-#[sea_orm(table_name = "page_revision")]
+#[sea_orm(table_name = "file_revision")]
 pub struct Model {
     #[sea_orm(primary_key)]
     pub revision_id: i64,
-    pub revision_type: PageRevisionType,
+    pub revision_type: Option<FileRevisionType>,
     pub created_at: DateTimeWithTimeZone,
     pub revision_number: i32,
-    pub page_id: i64,
-    pub site_id: i64,
-    pub user_id: i64,
-    pub changes: Json,
-    pub wikitext_hash: Vec<u8>,
-    pub compiled_hash: Vec<u8>,
-    pub compiled_at: DateTimeWithTimeZone,
     #[sea_orm(column_type = "Text")]
-    pub compiled_generator: String,
+    pub file_id: String,
+    pub page_id: i64,
+    pub user_id: i64,
+    #[sea_orm(column_type = "Text")]
+    pub name: String,
+    pub s3_hash: Vec<u8>,
+    #[sea_orm(column_type = "Text")]
+    pub mime_hint: String,
+    pub size_hint: i64,
+    pub licensing: Json,
+    pub changes: Json,
     #[sea_orm(column_type = "Text")]
     pub comments: String,
     pub hidden: Json,
-    #[sea_orm(column_type = "Text")]
-    pub title: String,
-    #[sea_orm(column_type = "Text", nullable)]
-    pub alt_title: Option<String>,
-    #[sea_orm(column_type = "Text")]
-    pub slug: String,
-    pub tags: Json,
 }
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
 pub enum Relation {
     #[sea_orm(
-        belongs_to = "super::text::Entity",
-        from = "Column::CompiledHash",
-        to = "super::text::Column::Hash",
+        belongs_to = "super::file::Entity",
+        from = "Column::FileId",
+        to = "super::file::Column::FileId",
         on_update = "NoAction",
         on_delete = "NoAction"
     )]
-    Text2,
+    File,
     #[sea_orm(
         belongs_to = "super::page::Entity",
         from = "Column::PageId",
@@ -52,14 +48,6 @@ pub enum Relation {
     )]
     Page,
     #[sea_orm(
-        belongs_to = "super::site::Entity",
-        from = "Column::SiteId",
-        to = "super::site::Column::SiteId",
-        on_update = "NoAction",
-        on_delete = "NoAction"
-    )]
-    Site,
-    #[sea_orm(
         belongs_to = "super::users::Entity",
         from = "Column::UserId",
         to = "super::users::Column::Id",
@@ -67,25 +55,17 @@ pub enum Relation {
         on_delete = "NoAction"
     )]
     Users,
-    #[sea_orm(
-        belongs_to = "super::text::Entity",
-        from = "Column::WikitextHash",
-        to = "super::text::Column::Hash",
-        on_update = "NoAction",
-        on_delete = "NoAction"
-    )]
-    Text1,
+}
+
+impl Related<super::file::Entity> for Entity {
+    fn to() -> RelationDef {
+        Relation::File.def()
+    }
 }
 
 impl Related<super::page::Entity> for Entity {
     fn to() -> RelationDef {
         Relation::Page.def()
-    }
-}
-
-impl Related<super::site::Entity> for Entity {
-    fn to() -> RelationDef {
-        Relation::Site.def()
     }
 }
