+++ deepwell/src/models/prelude.rs
@@ -1,6 +1,7 @@
 //! SeaORM Entity. Generated by sea-orm-codegen 0.6.0
 
 pub use super::file::Entity as File;
+pub use super::file_revision::Entity as FileRevision;
 pub use super::forum_category::Entity as ForumCategory;
 pub use super::forum_group::Entity as ForumGroup;
 pub use super::forum_post::Entity as ForumPost;
