+++ deepwell/src/models/users.rs
@@ -40,8 +40,8 @@ pub struct Model {
 pub enum Relation {
     #[sea_orm(has_many = "super::page_revision::Entity")]
     PageRevision,
-    #[sea_orm(has_many = "super::file::Entity")]
-    File,
+    #[sea_orm(has_many = "super::file_revision::Entity")]
+    FileRevision,
 }
 
 impl Related<super::page_revision::Entity> for Entity {
@@ -50,9 +50,9 @@ impl Related<super::page_revision::Entity> for Entity {
     }
 }
 
-impl Related<super::file::Entity> for Entity {
+impl Related<super::file_revision::Entity> for Entity {
     fn to() -> RelationDef {
-        Relation::File.def()
+        Relation::FileRevision.def()
     }
 }
